feat: refactor imports to use Function directory for ApiRequest
